requirements.txt
# Vehicle Lights Pipeline - Pinned Dependencies
# Python 3.10 compatible

# Core dependencies (PyTorch installed separately via setup.sh)
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0

# Deep learning
# torch, torchvision, torchaudio installed via setup.sh with CUDA 11.8
# mmcv, mmengine, mmdet installed via mim in setup.sh

# Tracking
deep-sort-realtime==1.3.2
filterpy==1.4.5
scipy==1.11.4

# Data processing
PyYAML==6.0.1
pandas==2.0.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
colorama==0.4.6

# Video processing
imageio==2.33.0
imageio-ffmpeg==0.4.9

# Logging
loguru==0.7.2
